backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
email-validator==2.2.0
python-multipart==0.0.12
alembic==1.14.0
pytest==8.3.3
httpx==0.27.2
